requirements.txt
absl-py==2.1.0
accessible-pygments==0.0.4
alabaster==0.7.13
albumentations==0.5.2
astunparse==1.6.3
babel==2.16.0
beautifulsoup4==4.12.3
bunch==1.0.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
d2py==0.5.9
dataclasses==0.6
docutils==0.20.1
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
future==1.0.0
gast==0.4.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.8
imageio==2.35.1
imagesize==1.4.1
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
loguru==0.5.3
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.3.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.2
onnxruntime==1.19.2
opencv-python==4.4.0.46
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
Pillow==9.0.1
protobuf==4.25.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pydata-sphinx-theme==0.14.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ruamel.base==1.0.0
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.0.2
scipy==1.10.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
segmentation==0.2.2
shapely==2.0.6
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sympy==1.13.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.11
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
torchstat==0.0.7
torchvision==0.19.1
tqdm==4.61.1
transformers==4.45.1
triton==3.0.0
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
yacs==0.1.8
zipp==3.20.1
